Sender Firebase Auth Vesrion 1.0
