+++ config.json
@@ -1,0 +1,49 @@
+{
+    
+"ProjectId":
+    {
+    "apiKey": "AIzaSyDtQPBX8qQZ22xyNCaAANg02XDM_bfN0uk",
+    "authDomain": "add-user-e98e9.firebaseapp.com",
+    "projectId": "add-user-e98e9",
+    "databaseURL": "https://add-user-e98e9-default-rtdb.firebaseio.com",
+    "storageBucket": "add-user-e98e9.appspot.com",
+    "messagingSenderId": "858288259676",
+    "appId": "1:858288259676:web:52354ec6f2a90f13581c99",
+    "measurementId": "G-D0WPTFDN0K"
+    },
+
+"Project_1": {
+    "apiKey": "AIzaSyDtQPBX8qQZ22xyNCaAANg02XDM_bfN0uk",
+    "authDomain": "add-user-e98e9.firebaseapp.com",
+    "projectId": "add-user-e98e9",
+    "databaseURL": "https://add-user-e98e9-default-rtdb.firebaseio.com",
+    "storageBucket": "add-user-e98e9.appspot.com",
+    "messagingSenderId": "858288259676",
+    "appId": "1:858288259676:web:52354ec6f2a90f13581c99",
+    "measurementId": "G-D0WPTFDN0K"
+}, 
+
+"Project_2": {
+    "apiKey": "AIzaSyDtQPBX8qQZ22xyNCaAANg02XDM_bfN0uk",
+    "authDomain": "add-user-e98e9.firebaseapp.com",
+    "projectId": "add-user-e98e9",
+    "databaseURL": "https://add-user-e98e9-default-rtdb.firebaseio.com",
+    "storageBucket": "add-user-e98e9.appspot.com",
+    "messagingSenderId": "858288259676",
+    "appId": "1:858288259676:web:52354ec6f2a90f13581c99",
+    "measurementId": "G-D0WPTFDN0K"
+},
+
+"project_3": 
+{
+    "apiKey": "AIzaSyDtQPBX8qQZ22xyNCaAANg02XDM_bfN0uk",
+    "authDomain": "add-user-e98e9.firebaseapp.com",
+    "projectId": "add-user-e98e9",
+    "databaseURL": "https://add-user-e98e9-default-rtdb.firebaseio.com",
+    "storageBucket": "add-user-e98e9.appspot.com",
+    "messagingSenderId": "858288259676",
+    "appId": "1:858288259676:web:52354ec6f2a90f13581c99",
+    "measurementId": "G-D0WPTFDN0K"
+}
+
+}
